refactor(cli): remove V4 resource/capability commands
